Change grid_size back to uint due to many castings in the client plugin

+++ server/src/core/livegame.rs
@@ -1281,31 +1281,31 @@ impl LiveMatch {
             let cell_id = tile.cell_id as i32;
             if let Some(winning_team) = team {
                 let tile_up;
-                if cell_id - (self.config.grid_size) < 0 {
+                if cell_id - (self.config.grid_size as i32) < 0 {
                     tile_up = cell_id;
                 } else {
-                    tile_up = cell_id - self.config.grid_size;
+                    tile_up = cell_id - self.config.grid_size as i32;
                 }
 
                 let tile_left;
-                if (cell_id + 1) % self.config.grid_size == 1 {
+                if (cell_id + 1) % self.config.grid_size as i32 == 1 {
                     tile_left = cell_id;
                 } else {
                     tile_left = cell_id - 1;
                 }
 
                 let tile_right;
-                if (cell_id + 1) % self.config.grid_size == 0 {
+                if (cell_id + 1) % self.config.grid_size as i32 == 0 {
                     tile_right = cell_id;
                 } else {
                     tile_right = cell_id + 1;
                 }
 
                 let tile_down;
-                if cell_id + self.config.grid_size > self.cell_count() as i32 {
+                if cell_id + self.config.grid_size as i32 > self.cell_count() as i32 {
                     tile_down = cell_id;
                 } else {
-                    tile_down = cell_id + self.config.grid_size;
+                    tile_down = cell_id + self.config.grid_size as i32;
                 }
 
                 if tile_up >= 0 {
